feat: improved and fixed session manager
