+++ src/components/UserInfo.tsx
@@ -5,19 +5,16 @@ import { LogIn } from 'lucide-react';
 import { signOut } from '@/lib/actions';
 import { useSession } from '@/context/SessionProvider';
 import Image from 'next/image';
-import {useEffect} from "react";
-import {useSessionManager} from "@/helpers/hooks";
+import { useEffect, useState } from 'react';
+import { User } from '@supabase/supabase-js';
 
 type UserProfileProps = {
   toggleAction?: () => void;
 };
 
 export default function UserProfile({ toggleAction }: UserProfileProps) {
-  const { session : { user } } = useSessionManager();
+  const { user } = useSession();
 
-  useEffect(() => {
-    console.log("user", user);
-  }, [user]);
   const handleLogout = async () => {
     const response = await signOut();
