+++ src/helpers/hooks.ts
@@ -2,7 +2,7 @@ import { jwtDecode } from 'jwt-decode';
 import { exchangeCodeForSession, refreshToken, getSessionUser } from '@/lib/actions';
 import { useEffect, useRef, useState } from 'react';
 import { User } from '@supabase/auth-js/src/lib/types';
-import { UserSession } from "@/context/SessionProvider";
+import { UserSession } from '@/context/SessionProvider';
 
 type CallbackFunction<T extends any[]> = (...args: T) => void;
 
@@ -42,115 +42,4 @@ export function useDebouncedFunction<T extends any[]>(callback: CallbackFunction
   }, []);
 
   return debouncedFunction;
-}
-
-export function useSessionManager() {
-  const [isRefreshing, setIsRefreshing] = useState(false);
-  const [session, setSession] = useState<UserSession>({ user: null });
-  // const alpaca= useState<UserSession>({ user: null })
-  // const session = alpaca[0];
-  // const setSession = (userSession: UserSession) => {
-  //   console.log("The god shake", userSession);
-  //   alpaca[1](userSession);
-  // }
-
-  const getTokens = () => {
-    return {
-      accessToken: localStorage.getItem('accessToken'),
-      refreshToken: localStorage.getItem('refreshToken'),
-    };
-  };
-
-  const isTokenExpired = (token: string | null) => {
-    if (!token) return true;
-
-    try {
-      const decoded = jwtDecode<User & { exp: number }>(token);
-      // Check if token will expire in the next 15 minutes
-      const expirationBuffer = 15 * 60; // 15 minutes in seconds
-      return decoded.exp < Date.now() / 1000 + expirationBuffer;
-    } catch {
-      return true;
-    }
-  };
-
-  const refreshTokenIfNeeded = async () => {
-    const { accessToken, refreshToken: currentRefreshToken } = getTokens();
-
-    if (!isRefreshing && isTokenExpired(accessToken) && currentRefreshToken) {
-      setIsRefreshing(true);
-
-      try {
-        const response = await refreshToken(currentRefreshToken);
-
-        if (response.data?.session) {
-          localStorage.setItem('accessToken', response.data.session.access_token);
-          localStorage.setItem('refreshToken', response.data.session.refresh_token);
-debugger
-          setSession(response.data);
-
-          return true;
-        }
-      } catch (error) {
-        // Clear tokens if refresh fails
-        localStorage.removeItem('accessToken');
-        localStorage.removeItem('refreshToken');
-        return false;
-      } finally {
-        setIsRefreshing(false);
-      }
-    }
-    return !isTokenExpired(accessToken);
-  };
-
-  useEffect(() => {
-    const fetchSession = async () => {
-      const { data } = await getSessionUser();
-
-      if (data !== null) {
-        debugger
-        setSession(data);
-      }
-    };
-    fetchSession();
-  }, []);
-  
-  useEffect(() => {
-    refreshTokenIfNeeded();
-
-    const checkInterval = setInterval(refreshTokenIfNeeded, 60 * 1000); // Check every minute
-    return () => clearInterval(checkInterval);
-  }, []);
-
-  useEffect(() => {
-    const handleOAuthCallback = async () => {
-      const { accessToken } = getTokens();
-      const url = new URL(window.location.href);
-      const code = url.searchParams.get('code');
-
-      if (code && isTokenExpired(accessToken)) {
-        const session = await exchangeCodeForSession(code);
-
-        const accessToken = session.data.session?.access_token;
-        const refreshToken = session.data.session?.refresh_token;
-
-        if (accessToken && refreshToken) {
-          localStorage.setItem('accessToken', accessToken);
-          localStorage.setItem('refreshToken', refreshToken);
-  debugger
-          setSession(session.data);
-        }
-      }
-    };
-
-    handleOAuthCallback();
-  }, []);
-
-  return {
-    session,
-    setSession,
-    getTokens,
-    isTokenExpired,
-    refreshTokenIfNeeded,
-  };
 }
